SAN-279: Use method return value to fix sonar error

+++ src/main/java/uk/gov/companieshouse/api/testdata/service/impl/AccountPenaltiesServiceImpl.java
@@ -3,6 +3,7 @@ package uk.gov.companieshouse.api.testdata.service.impl;
 import java.util.List;
 import java.util.Objects;
 
+import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
@@ -82,8 +83,12 @@ public class AccountPenaltiesServiceImpl implements AccountPenaltiesService {
     @Override
     public ResponseEntity<Void> deleteAccountPenalties(String companyCode, String customerCode)
             throws NoDataFoundException {
-        repository.findAllByCompanyCodeAndCustomerCode(companyCode, customerCode)
-                        .orElseThrow(() -> new NoDataFoundException("no account penalties"));
+        Optional<AccountPenalties> accountPenalties =
+                repository.findAllByCompanyCodeAndCustomerCode(companyCode, customerCode);
+
+        if (accountPenalties.isEmpty()) {
+            throw new NoDataFoundException("no account penalties");
+        }
 
         repository.deleteByCompanyCodeAndCustomerCode(companyCode, customerCode);
